week4/mqpub/requirements.txt
pyzmq==26.2.1
paho-mqtt==2.1.0
